refactor: rename log utility file to ai-log
- Rename `src/utils/log.ts` to `src/utils/ai-log.ts`.
- Update imports in commands and AI service files to reflect the new file name.
- Use `AICOMMIT_LOGS_DIR` from config in the `log` command for directory path consistency.
