+++ src/services/ai/anthropic.service.ts
@@ -5,8 +5,8 @@ import { Observable, catchError, concatMap, from, map, of } from 'rxjs';
 import { fromPromise } from 'rxjs/internal/observable/innerFrom';
 
 import { AIResponse, AIService, AIServiceError, AIServiceParams } from './ai.service.js';
+import { RequestType, createLogResponse } from '../../utils/ai-log.js';
 import { KnownError } from '../../utils/error.js';
-import { RequestType, createLogResponse } from '../../utils/log.js';
 import { DEFAULT_PROMPT_OPTIONS, PromptOptions, codeReviewPrompt, generatePrompt } from '../../utils/prompt.js';
 
 export interface AnthropicServiceError extends AIServiceError {
